TST: remove duplicated index tests and put in one file (#27045)
